requirements.txt
# Existing Repository Requirements
langchain==0.3.25
langchain_community==0.3.24
openai==1.78.1
pydantic>=2.11.4
pytest>=8.3.5
sentence-transformers>=4.1.0
transformers>=4.51.3
chromadb==1.0.9
jupyter==1.1.1
tiktoken==0.9.0
python_dotenv>=1.1.0
lancedb==0.22.0